Изменение : MPEGSplitter - корректировка чтения PSI, необходимо учитывать CRC32 в конце данных и не учитывать их при обработке таблиц.
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@850 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/filters/parser/BaseSplitter/BaseSplitterFileEx.h
@@ -347,10 +347,10 @@ public:
 
 	struct trsechdr
 	{
-		WORD section_syntax_indicator:1;
-		WORD zero:1;
-		WORD reserved1:2;
-		WORD section_length:12;
+		BYTE section_syntax_indicator:1;
+		BYTE zero:1;
+		BYTE reserved1:2;
+		int section_length:12;
 		WORD transport_stream_id;
 		BYTE table_id;
 		BYTE reserved2:2;
